pyproject.toml
[project]
name = "structured-output-benchmark"
version = "0.1.0"
description = "Benchmark various structured output frameworks with identical inputs"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.131.0",
    "guardrails-ai>=0.9.0",
    "instructor>=1.14.5",
    "langchain-openai>=1.1.10",
    "llama-index-llms-openai-like>=0.1.3",
    "llama-index-program-openai>=0.1.7",
    "loguru>=0.7.3",
    "marvin>=3.2.6",
    "matplotlib>=3.10.8",
    "mirascope[openai]>=2.2.2",
    "openai>=2.21.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "pyyaml>=6.0.3",
    "streamlit>=1.55.0",
    "uvicorn[standard]>=0.41.0",
]
